fix intel warning

+++ src/ref_split.c
@@ -224,7 +224,7 @@ REF_STATUS ref_split_pass(REF_GRID ref_grid) {
                                      ref_edge_e2n(ref_edge, 1, edge), new_node),
           "form edge split cav");
       if (transcript)
-        printf("try cavity status %d\n", ref_cavity_state(ref_cavity));
+        printf("try cavity status %d\n", (int)ref_cavity_state(ref_cavity));
       if (REF_SUCCESS != ref_cavity_enlarge_combined(ref_cavity))
         REF_WHERE("enlarge"); /* note but skip cavity failures */
       if (REF_CAVITY_VISIBLE == ref_cavity_state(ref_cavity)) {
@@ -254,7 +254,7 @@ REF_STATUS ref_split_pass(REF_GRID ref_grid) {
         }
       } else {
         if (transcript)
-          printf("cavity not visible %d\n", ref_cavity_state(ref_cavity));
+          printf("cavity not visible %d\n", (int)ref_cavity_state(ref_cavity));
       }
       if (REF_CAVITY_PARTITION_CONSTRAINED == ref_cavity_state(ref_cavity)) {
         if (span_parts) RSS(ref_list_push(para_cavity, edge), "push");
